requirements.txt
Django==2.1.1
djangorestframework==3.8.2
requests==2.19.1
django-rest-auth==0.9.3
django-extensions==2.1.2